+++ react/StorefrontUserWidget.tsx
@@ -1,0 +1,3 @@
+import StorefrontUserWidget from './components/UserWidget'
+
+export default StorefrontUserWidget