+++ react/components/CostCenterDetails.tsx
@@ -1,6 +1,6 @@
 import type { FunctionComponent } from 'react'
 import React, { useEffect, useState, useContext } from 'react'
-import { useQuery, useMutation } from 'react-apollo'
+import { useQuery, useMutation, useLazyQuery } from 'react-apollo'
 import {
   Layout,
   PageHeader,
@@ -11,10 +11,10 @@ import {
   Button,
   Spinner,
   Input,
+  Toggle,
   ToastContext,
 } from 'vtex.styleguide'
 import { useIntl, FormattedMessage, defineMessages } from 'react-intl'
-import { useRuntime } from 'vtex.render-runtime'
 import { AddressRules, AddressSummary } from 'vtex.address-form'
 
 import storageFactory from '../utils/storage'
@@ -24,12 +24,32 @@ import EditAddressModal from './EditAddressModal'
 import DeleteAddressModal from './DeleteAddressModal'
 import DeleteCostCenterModal from './DeleteCostCenterModal'
 import GET_COST_CENTER from '../graphql/getCostCenterStorefront.graphql'
+import GET_ORGANIZATION from '../graphql/getOrganizationStorefront.graphql'
 import UPDATE_COST_CENTER from '../graphql/updateCostCenter.graphql'
 import DELETE_COST_CENTER from '../graphql/deleteCostCenter.graphql'
+import GET_PERMISSIONS from '../graphql/getPermissions.graphql'
+
+interface RouterProps {
+  match: Match
+  history: any
+}
+
+interface Match {
+  isExact: boolean
+  params: any
+  path: string
+  url: string
+}
+
+interface PaymentTerm {
+  id: string
+  name: string
+}
 
 const localStore = storageFactory(() => localStorage)
 let isAuthenticated =
-  JSON.parse(String(localStore.getItem('orderquote_isAuthenticated'))) ?? false
+  JSON.parse(String(localStore.getItem('b2b-organizations_isAuthenticated'))) ??
+  false
 
 const storePrefix = 'store/b2b-organizations.'
 
@@ -61,15 +81,22 @@ const messages = defineMessages({
   addresses: {
     id: `${storePrefix}costCenter-details.addresses`,
   },
+  paymentTerms: {
+    id: `${storePrefix}costCenter-details.payment-terms`,
+  },
+  addressesSubtitle: {
+    id: `${storePrefix}costCenter-details.addresses.helpText`,
+  },
+  paymentTermsSubtitle: {
+    id: `${storePrefix}costCenter-details.payment-terms.helpText`,
+  },
 })
 
-const CostCenterDetails: FunctionComponent = () => {
+const CostCenterDetails: FunctionComponent<RouterProps> = ({
+  match: { params },
+  history,
+}) => {
   const { formatMessage } = useIntl()
-
-  const {
-    route: { params },
-    navigate,
-  } = useRuntime()
 
   const sessionResponse: any = useSessionResponse()
 
@@ -78,7 +105,7 @@ const CostCenterDetails: FunctionComponent = () => {
       sessionResponse?.namespaces?.profile?.isAuthenticated?.value === 'true'
 
     localStore.setItem(
-      'orderquote_isAuthenticated',
+      'b2b-organizations_isAuthenticated',
       JSON.stringify(isAuthenticated)
     )
   }
@@ -93,9 +120,15 @@ const CostCenterDetails: FunctionComponent = () => {
     showToast({ message: translatedMessage, duration: 5000, action })
   }
 
+  const [permissionsState, setPermissionsState] = useState([] as string[])
   const [loadingState, setLoadingState] = useState(false)
   const [costCenterName, setCostCenterName] = useState('')
   const [addresses, setAddresses] = useState([] as Address[])
+  const [paymentTerms, setPaymentTerms] = useState([] as PaymentTerm[])
+  const [paymentTermOptions, setPaymentTermOptions] = useState(
+    [] as PaymentTerm[]
+  )
+
   const [newAddressModalState, setNewAddressModalState] = useState({
     isOpen: false,
   })
@@ -120,15 +153,53 @@ const CostCenterDetails: FunctionComponent = () => {
     ssr: false,
   })
 
+  const [getOrganization, { data: organizationData }] = useLazyQuery(
+    GET_ORGANIZATION
+  )
+
+  const { data: permissionsData } = useQuery(GET_PERMISSIONS, { ssr: false })
+
   const [updateCostCenter] = useMutation(UPDATE_COST_CENTER)
   const [deleteCostCenter] = useMutation(DELETE_COST_CENTER)
 
   useEffect(() => {
-    if (addresses.length || !data?.getCostCenterById?.addresses?.length) return
+    if (!data?.getCostCenterByIdStorefront) return
 
-    setCostCenterName(data.getCostCenterById.name)
-    setAddresses(data.getCostCenterById.addresses)
+    setCostCenterName(data.getCostCenterByIdStorefront.name)
+    setAddresses(data.getCostCenterByIdStorefront.addresses)
+    setPaymentTerms(
+      data?.getCostCenterByIdStorefront?.paymentTerms?.length
+        ? data?.getCostCenterByIdStorefront?.paymentTerms
+        : []
+    )
+    getOrganization({
+      variables: { id: data.getCostCenterByIdStorefront.organization },
+    })
   }, [data])
+
+  useEffect(() => {
+    const termOptions = organizationData?.getOrganizationByIdStorefront
+      ?.paymentTerms?.length
+      ? organizationData.getOrganizationByIdStorefront.paymentTerms
+      : []
+
+    setPaymentTermOptions(termOptions)
+
+    // enable all available payment terms by default
+    if (!paymentTerms.length) {
+      setPaymentTerms(termOptions)
+    }
+  }, [organizationData])
+
+  useEffect(() => {
+    if (!permissionsData) return
+
+    const { permissions = [] } = permissionsData.checkUserPermission ?? {}
+
+    if (permissions.length) {
+      setPermissionsState(permissions)
+    }
+  }, [permissionsData])
 
   const handleUpdateCostCenter = () => {
     setLoadingState(true)
@@ -137,6 +208,7 @@ const CostCenterDetails: FunctionComponent = () => {
       input: {
         name: costCenterName,
         addresses,
+        paymentTerms,
       },
     }
 
@@ -157,10 +229,9 @@ const CostCenterDetails: FunctionComponent = () => {
     setLoadingState(true)
     deleteCostCenter({ variables: { id: params?.id } })
       .then(() => {
-        navigate({
-          page: 'store.organization-details',
-          params: { id: data.getCostCenterById.organization },
-        })
+        history.push(
+          `/organization/${data.getCostCenterByIdStorefront.organization}`
+        )
       })
       .catch(error => {
         console.error(error)
@@ -261,6 +332,22 @@ const CostCenterDetails: FunctionComponent = () => {
     handleCloseModals()
   }
 
+  const handleTogglePaymentTerm = (id: string) => {
+    let newTerms = paymentTerms
+    const termOption = paymentTermOptions.find(term => term.id === id)
+
+    if (!termOption) return
+    const enabled = paymentTerms.find(term => term.id === id)
+
+    if (enabled) {
+      newTerms = paymentTerms.filter(term => term.id !== enabled.id)
+    } else {
+      newTerms.push(termOption)
+    }
+
+    setPaymentTerms(newTerms)
+  }
+
   const options = (addressId: string) => [
     {
       label: formatMessage(messages.addressEdit),
@@ -281,9 +368,7 @@ const CostCenterDetails: FunctionComponent = () => {
             title={formatMessage(messages.pageTitle)}
             linkLabel={formatMessage(messages.back)}
             onLinkClick={() => {
-              navigate({
-                page: 'store.organization-details',
-              })
+              history.push(`/organization`)
             }}
           />
         }
@@ -304,9 +389,7 @@ const CostCenterDetails: FunctionComponent = () => {
             title={formatMessage(messages.pageTitle)}
             linkLabel={formatMessage(messages.back)}
             onLinkClick={() => {
-              navigate({
-                page: 'store.organization-details',
-              })
+              history.push(`/organization`)
             }}
           />
         }
@@ -328,19 +411,26 @@ const CostCenterDetails: FunctionComponent = () => {
       pageHeader={
         <PageHeader
           title={formatMessage(messages.pageTitle)}
-          linkLabel={formatMessage(messages.back)}
+          linkLabel={
+            organizationData?.getOrganizationByIdStorefront?.name ??
+            formatMessage(messages.back)
+          }
           onLinkClick={() => {
-            navigate({
-              page: 'store.organization-details',
-              params: { id: data.getCostCenterById.organization },
-            })
+            history.push(
+              `/organization/${data.getCostCenterByIdStorefront.organization}`
+            )
           }}
         >
           <span className="mr4">
             <Button
               variation="primary"
               isLoading={loadingState}
-              disabled={!costCenterName || !addresses.length}
+              disabled={
+                !costCenterName ||
+                !addresses.length ||
+                (paymentTermOptions.length > 0 && paymentTerms.length === 0) ||
+                !permissionsState.includes('create-cost-center-organization')
+              }
               onClick={() => handleUpdateCostCenter()}
             >
               <FormattedMessage id="store/b2b-organizations.costCenter-details.button.save" />
@@ -350,6 +440,9 @@ const CostCenterDetails: FunctionComponent = () => {
             variation="danger"
             isLoading={loadingState}
             onClick={() => handleDeleteCostCenterModal()}
+            disabled={
+              !permissionsState.includes('create-cost-center-organization')
+            }
           >
             <FormattedMessage id="store/b2b-organizations.costCenter-details.button.delete" />
           </Button>
@@ -364,10 +457,42 @@ const CostCenterDetails: FunctionComponent = () => {
           onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
             setCostCenterName(e.target.value)
           }}
+          readOnly={
+            !permissionsState.includes('create-cost-center-organization')
+          }
           required
         />
       </PageBlock>
-      <PageBlock title={formatMessage(messages.addresses)}>
+      {paymentTermOptions.length > 0 && (
+        <PageBlock
+          title={formatMessage(messages.paymentTerms)}
+          subtitle={formatMessage(messages.paymentTermsSubtitle)}
+        >
+          {paymentTermOptions.map((option, index) => {
+            const checked = paymentTerms.some(term => term.id === option.id)
+
+            return (
+              <div key={index} className="mv4">
+                <Toggle
+                  label={option.name}
+                  semantic
+                  checked={checked}
+                  onChange={() => handleTogglePaymentTerm(option.id)}
+                  disabled={
+                    !permissionsState.includes(
+                      'create-cost-center-organization'
+                    )
+                  }
+                ></Toggle>
+              </div>
+            )
+          })}
+        </PageBlock>
+      )}
+      <PageBlock
+        title={formatMessage(messages.addresses)}
+        subtitle={formatMessage(messages.addressesSubtitle)}
+      >
         <div className="flex">
           {addresses.map((address: any, index) => {
             return (
@@ -384,31 +509,37 @@ const CostCenterDetails: FunctionComponent = () => {
                       </AddressRules>
                     </div>
                     <div>
-                      <ActionMenu
-                        buttonProps={{
-                          variation: 'tertiary',
-                          icon: <IconOptionsDots color="currentColor" />,
-                        }}
-                        options={options(address.addressId)}
-                      />
+                      {permissionsState.includes(
+                        'create-cost-center-organization'
+                      ) && (
+                        <ActionMenu
+                          buttonProps={{
+                            variation: 'tertiary',
+                            icon: <IconOptionsDots color="currentColor" />,
+                          }}
+                          options={options(address.addressId)}
+                        />
+                      )}
                     </div>
                   </div>
                 </Card>
               </div>
             )
           })}
-          <div className="w-25 ma3">
-            <Card>
-              <div className="flex justify-center">
-                <Button
-                  variation="primary"
-                  onClick={() => handleNewAddressModal()}
-                >
-                  <FormattedMessage id="store/b2b-organizations.costCenter-details.address.new" />
-                </Button>
-              </div>
-            </Card>
-          </div>
+          {permissionsState.includes('create-cost-center-organization') && (
+            <div className="w-25 ma3">
+              <Card>
+                <div className="flex justify-center">
+                  <Button
+                    variation="primary"
+                    onClick={() => handleNewAddressModal()}
+                  >
+                    <FormattedMessage id="store/b2b-organizations.costCenter-details.address.new" />
+                  </Button>
+                </div>
+              </Card>
+            </div>
+          )}
         </div>
       </PageBlock>